Remove unsupported dynamic variables
Some of the dynamic variables in this doc are no longer supported. Found [this postman issue request](https://github.com/postmanlabs/postman-app-support/issues/7868) earlier today. 

There might be some I missed and variables that need to be added, but these were the ones I found to be no longer working.